requirements.txt
Package                          Version
-------------------------------- -------
annotated-types                  0.7.0
asgiref                          3.9.1
crispy-bootstrap5                2025.6
Django                           5.2.6
django-bootstrap-datepicker-plus 5.0.6
django-crispy-forms              2.4
pip                              25.2
pydantic                         2.12.3
pydantic_core                    2.41.4
setuptools                       65.5.0
sqlparse                         0.5.3
typing_extensions                4.15.0
typing-inspection                0.4.2
pip
